Model as protocol removed. CGFloat extension.
