+++ RotaryDialPhone/RotaryDialView.swift
@@ -8,14 +8,40 @@
 
 import UIKit
 
-class RotaryDialView: UIImageView, RotaryDialModel {
+class RotaryDialView: UIImageView {
     var path: UIBezierPath!
     var numberFontSize: CGFloat!
     
+    let holesCount = 10
     var holesRadius: CGFloat!
     var distanceFromHolesToCenter: CGFloat!
     var holesSeparationAngle: CGFloat!
     var firstHoleAngle: CGFloat!
+    
+    var initHoleAngle: CGFloat {
+        return firstHoleAngle - holesSeparationAngle / 2.0
+    }
+    
+    var lockAngle: CGFloat {
+        return CGFloat.M_2_PI - holesSeparationAngle / 2.0
+    }
+    
+    func hole(_ index: Int) -> CGPoint {
+        let angle = firstHoleAngle + holesSeparationAngle * CGFloat(index)
+        
+        return CGPoint(
+            x: distanceFromHolesToCenter * cos(angle) + bounds.midX,
+            y: distanceFromHolesToCenter * sin(angle) + bounds.midY
+        )
+    }
+    
+    func number(_ index: Int) -> Int {
+        if index > 0 {
+            return holesCount - index
+        }
+        
+        return 0
+    }
 }
 
 extension RotaryDialView: CirclePath {
